Fix version extraction from NuGet package DLL path
This change fixes the extraction of a package version from the filepath
to the DLL file of the NuGet package.

The prior implementation assumed the name of the package
(i.e., xunit.core) would always be found in the path. However, that
example is located at 'xunit.extensibility.core'. So when removing the
prefix of NuGet local package directory + package name the replace
failed. This resulted in the drive letter being returned as it was
before the first slash (/).

The new implementation splits the path on the directory separator, and
extracts the 4th last segment. This assumes the directory structure of
NuGet packages remains consistent with:
'C:/<local-nuget-folder>/<id>/<version>/lib/<target-framework>/<dll>'.

Issue: #7